Split Vim9 into compile and command crates
