+++ rust_vim9cmds/src/lib.rs
@@ -2,39 +2,12 @@ use std::ffi::CStr;
 use std::os::raw::c_char;
 
 use rust_eval::{typval_T, ValUnion, Vartype};
-
-mod types;
-mod parser;
-mod compiler;
-mod executor;
-
-use parser::parse_line;
-use compiler::compile;
-use executor::execute;
-use types::Vim9Type;
-
-pub fn eval_expr(expr: &str) -> Option<i64> {
-    let ast = parse_line(expr)?;
-    let prog = compile(&ast);
-    if prog.result_type != Vim9Type::Number {
-        return None;
-    }
-    Some(execute(&prog))
-}
-
-pub fn eval_bool_expr(expr: &str) -> Option<bool> {
-    let ast = parse_line(expr)?;
-    let prog = compile(&ast);
-    if prog.result_type != Vim9Type::Bool {
-        return None;
-    }
-    Some(execute(&prog) != 0)
-}
+use rust_vim9compile::{eval_bool_expr, eval_expr};
 
 pub mod cmds {
     use super::*;
     pub fn execute(expr: &str, out: *mut typval_T) -> bool {
-        match super::eval_expr(expr) {
+        match eval_expr(expr) {
             Some(n) => unsafe {
                 (*out).v_type = Vartype::VAR_NUMBER;
                 (*out).vval = ValUnion { v_number: n };
@@ -100,41 +73,5 @@ pub extern "C" fn vim9_declare_error_rs(name: *const c_char) {
             _ => return,
         };
         eprintln!("{msg}");
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rust_eval::eval_expr_rs;
-
-    #[test]
-    fn eval_simple_add() {
-        let expr = "1 + 2 + 3";
-        assert_eq!(eval_expr(expr), Some(6));
-    }
-
-    #[test]
-    fn echo_command_returns_value() {
-        let expr = "echo 4 + 5";
-        assert_eq!(eval_expr(expr), Some(9));
-    }
-
-    #[test]
-    fn compatibility_with_rust_eval() {
-        let expr = "1 + 2 + 3";
-        let ours = eval_expr(expr).unwrap();
-        let c_expr = std::ffi::CString::new(expr).unwrap();
-        let mut tv = typval_T { v_type: Vartype::VAR_UNKNOWN, v_lock: 0, vval: ValUnion { v_number: 0 } };
-        let ok = eval_expr_rs(c_expr.as_ptr(), &mut tv as *mut typval_T);
-        assert!(ok);
-        let theirs = unsafe { tv.vval.v_number };
-        assert_eq!(ours, theirs);
-    }
-
-    #[test]
-    fn eval_bool_comparison() {
-        let expr = "1 < 2";
-        assert_eq!(eval_bool_expr(expr), Some(true));
     }
 }
