+++ rust_vim9compile/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "rust_vim9compile"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "rust_vim9compile"
+crate-type = ["rlib"]
+
+[dependencies]
